fixed the half ops issue and added some other code which wont do much atm

+++ src/hash.h
@@ -191,7 +191,7 @@ struct exemptNode {
 struct modeNode {
     struct chanNode *channel;
     struct userNode *user;
-    unsigned short modes;
+    long modes;
     short oplevel;
     time_t idle_since;
 };
